MAINT: Refactor for mne-qt-browser
